Factor out building the keyset filter predicate
